third_party/blink/web_tests/platform/generic/http/tests/websocket/handshake-error-expected.txt

Handshake error test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS onerror() was called
Closed.
PASS ws.readyState is 3
PASS successfullyParsed is true

TEST COMPLETE
